Fix potential problem with integer calculations.
Parameters were specified as integers so internal calculations are
performed only with integers while the result is a long integer.

git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@2401 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/particleContainer/LinkedCells.cpp
@@ -626,7 +626,7 @@ unsigned long int LinkedCells::getCellIndexOfMolecule(Molecule* molecule) const 
 	return this->cellIndexOf3DIndex( cellIndex[0], cellIndex[1], cellIndex[2] );
 }
 
-long int LinkedCells::cellIndexOf3DIndex(int xIndex, int yIndex, int zIndex) const {
+long int LinkedCells::cellIndexOf3DIndex(long int xIndex, long int yIndex, long int zIndex) const {
 	return (zIndex * _cellsPerDimension[1] + yIndex) * _cellsPerDimension[0] + xIndex;
 }
