+++ src/particleContainer/LinkedCells.h
@@ -262,7 +262,7 @@ private:
 	//! The method can also be used to get the offset between two cells in the cell
 	//! vector when called with the 3D cell index offsets (e.g. x: one cell to the left,
 	//! y: two cells back, z: one cell up,...)
-	long int cellIndexOf3DIndex(int xIndex, int yIndex, int zIndex) const;
+	long int cellIndexOf3DIndex(long int xIndex, long int yIndex, long int zIndex) const;
